requirements.txt
flask>=2.0.1
requests==2.31.0
python-dotenv==1.0.0
instagrapi==1.17.20
Pillow>=8.1.1
google-generativeai==0.3.2
selenium==4.18.1
webdriver-manager==4.0.1
pydantic==1.10.13